refactor: update storage location components imports to use shadcn directory
