+++ features/storage-locations/components/storage-location-alert-dialogs.tsx
@@ -8,8 +8,8 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
-import { Button } from '@/components/ui/button';
+} from '@/components/shadcn/alert-dialog';
+import { Button } from '@/components/shadcn/button';
 import { Loader2 } from 'lucide-react';
 
 // ---------------------------------------------------------------------------
